requirements-dev.txt
-r requirements.txt
coverage==6.5.0
pytest-django==4.5.2
gunicorn
black
